Correccion strdup: cambio en el type de int a size_t ya que ft_strlen devuelve ese tipo de dato

+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: fcarranz <fcarranz@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/14 13:35:51 by fcarranz          #+#    #+#             */
-/*   Updated: 2024/01/22 10:22:37 by fcarranz         ###   ########.fr       */
+/*   Updated: 2024/01/27 13:03:04 by fcarranz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strdup(const char *s1)
 {
-	int		len;
+	size_t	len;
 	char	*dst;
 
 	len = ft_strlen((char *)s1) + 1;
